Fix: Resolve build errors from feature implementation
This commit fixes several build errors that arose after the initial implementation of the comparison, sync, and versioning features.

- Resolved CS0111 (duplicate method) errors by consolidating all `MainForm` helper and operational logic. The new `MainForm.Methods.cs` was deleted, and its logic was merged into the pre-existing `MainForm.Operations.cs` file.
- Fixed CS0246 (type/namespace not found) errors in `VersionHistoryForm.cs` by adding the required `using S3FileManager.Services;` directive.
- Addressed CS8625 (nullable reference type) warning in `S3Service.cs` by modifying the `DownloadFileAsync` method to handle the optional `versionId` parameter more robustly.
