+++ AWSS3Sync.csproj
@@ -9,11 +9,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AWSSDK.Core" Version="3.7.500.12" />
-    <PackageReference Include="AWSSDK.S3" Version="3.7.500.12" />
-    <PackageReference Include="AWSSDK.CognitoIdentity" Version="3.7.501.3" />
-    <PackageReference Include="AWSSDK.CognitoIdentityProvider" Version="3.7.404" />
-    <PackageReference Include="Amazon.Extensions.CognitoAuthentication" Version="2.5.6" />
+    <PackageReference Include="AWSSDK.Core" Version="4.0.0.26" />
+    <PackageReference Include="AWSSDK.S3" Version="4.0.6.12" />
     <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="9.0.8" />
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.8" />
     <PackageReference Include="Microsoft.Extensions.FileProviders.Abstractions" Version="9.0.8" />
@@ -24,10 +21,8 @@
     <PackageReference Include="System.Memory" Version="4.6.3" />
     <PackageReference Include="System.Numerics.Vectors" Version="4.6.1" />
     <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="6.1.2" />
-    <PackageReference Include="System.Security.Cryptography.ProtectedData" Version="9.0.8" />
     <PackageReference Include="System.Text.Encodings.Web" Version="9.0.8" />
     <PackageReference Include="System.Text.Json" Version="9.0.8" />
-    <PackageReference Include="System.Threading.Channels" Version="9.0.8" />
     <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.6.3" />
     <PackageReference Include="System.ValueTuple" Version="4.6.1" />
   </ItemGroup>
