+++ Forms/VersionHistoryForm.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Windows.Forms;
 using S3FileManager.Models;
+using S3FileManager.Services;
 
 namespace S3FileManager.Forms
 {
