+++ Services/MetadataService.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Windows.Forms;
 using Amazon.S3;
 using Amazon.S3.Model;
 using S3FileManager.Models;
@@ -57,7 +58,7 @@ namespace S3FileManager.Services
             }
             catch (Exception ex)
             {
-                throw new InvalidOperationException($"An unexpected error occurred while retrieving file permissions for '{key}': {ex.Message}", ex);
+                MessageBox.Show($"An unexpected error occurred while retrieving file permissions for '{key}': {ex.Message}", "Permission Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
 
             // Default to Administrator only if no tags are found or an error occurs
@@ -212,9 +213,6 @@ namespace S3FileManager.Services
                 // Don't fail the main operation if auto-tagging fails, just log it
                 MessageBox.Show($"Warning: Could not auto-assign permission tag to '{key}': {ex.Message}",
                     "Auto-Tagging Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-
-                // Don't fail the main operation if auto-tagging fails, just throw a descriptive exception
-                // throw new InvalidOperationException($"Warning: Could not auto-assign permission tag to '{key}': {ex.Message}", ex);
             }
         }
